+++ gkraken/di.py
@@ -26,7 +26,8 @@ from peewee import SqliteDatabase
 from rx.disposables import CompositeDisposable
 from rx.subjects import Subject
 
-from gkraken.conf import APP_PACKAGE_NAME, APP_MAIN_UI_NAME, APP_DB_NAME, APP_EDIT_SPEED_PROFILE_UI_NAME
+from gkraken.conf import APP_PACKAGE_NAME, APP_MAIN_UI_NAME, APP_DB_NAME, APP_EDIT_SPEED_PROFILE_UI_NAME, \
+    APP_PREFERENCES_UI_NAME
 from gkraken.util.path import get_data_path, get_config_path
 
 LOG = logging.getLogger(__name__)
@@ -35,6 +36,7 @@ SpeedProfileChangedSubject = Key("SpeedProfileChangedSubject")
 SpeedStepChangedSubject = Key("SpeedStepChangedSubject")
 MainBuilder = Key(APP_MAIN_UI_NAME)
 EditSpeedProfileBuilder = Key(APP_EDIT_SPEED_PROFILE_UI_NAME)
+PreferencesBuilder = Key(APP_PREFERENCES_UI_NAME)
 
 
 # pylint: disable=no-self-use
@@ -55,6 +57,15 @@ class ProviderModule(Module):
         builder = Gtk.Builder()
         builder.set_translation_domain(APP_PACKAGE_NAME)
         builder.add_from_file(get_data_path(APP_EDIT_SPEED_PROFILE_UI_NAME))
+        return builder
+
+    @singleton
+    @provider
+    def provide_preferences_builder(self) -> PreferencesBuilder:
+        LOG.debug("provide Gtk.Builder")
+        builder = Gtk.Builder()
+        builder.set_translation_domain(APP_PACKAGE_NAME)
+        builder.add_from_file(get_data_path(APP_PREFERENCES_UI_NAME))
         return builder
 
     @singleton